+++ NEWS.md
@@ -1,5 +1,7 @@
 # gtsummary (development version)
 
+* Updates to account for changes in {forcats} v1.0.0 and {dplyr} v1.1.0.
+
 * Updated `style_ratio()` to now format negative values.
 * Bug fix in `add_ci.tbl_svysummary()` for dichotomous variables.